[JENKINS-56927] Pulled duplicates to common superclass
